BREAKS COMPAT! Converts to SL-compatible sit targets
An option is left to switch into the legacy mode for grids who wish to not break content.  This means that this setting can be set per-region, which allows for a range of possibilities.

The choice to default to SL-compat mode was made to facilitate new grids using this codebase: most scripts will be coming from SecondLife and it's really very annoying to have to fix them up - assuming the importer even knows how.