Await/async all the way down
Fixes all the compiler errors I noted in the previous commit.

Please note that this commit may cause more exceptions to crash the software as before they were masked away and hidden from sight by the way the methods were being called.  Now any exceptions will either fire at their callsite in the case of an async Task, or in the general exception handler for an async void.

In some cases I split off the async void event handler from the async task.  This is a violation of YAGNI, as it's for "future" unit testing...